+++ src/lib/util.js
@@ -3,13 +3,70 @@
 // helper to swap item to end and pop(), instead of splice()ing
 export function removeUnorderedListItem(list, item) {
     var i = list.indexOf(item)
-    if (i < 0) { return }
+    if (i < 0) return
     if (i === list.length - 1) {
         list.pop()
     } else {
         list[i] = list.pop()
     }
 }
+
+
+
+// loop over a function for a few ms, or until it returns true
+export function loopForTime(maxTimeInMS, callback, startTime) {
+    var cutoff = (startTime || performance.now()) + maxTimeInMS
+    var maxIter = 1000 // sanity check
+    for (var i = 0; i < maxIter; i++) {
+        var res = callback()
+        if (res || performance.now() > cutoff) return
+    }
+}
+
+
+
+
+
+// ....
+export function numberOfVoxelsInSphere(rad) {
+    if (rad === prevRad) return prevAnswer
+    var ext = Math.ceil(rad), ct = 0, rsq = rad * rad
+    for (var i = -ext; i <= ext; ++i) {
+        for (var j = -ext; j <= ext; ++j) {
+            for (var k = -ext; k <= ext; ++k) {
+                var dsq = i * i + j * j + k * k
+                if (dsq < rsq) ct++
+            }
+        }
+    }
+    prevRad = rad
+    prevAnswer = ct
+    return ct
+}
+var prevRad = 0, prevAnswer = 0
+
+
+
+
+// sorts [A, B, C] and [3, 1, 2] into [B, C, A]
+export function sortByReferenceArray(array, ref) {
+    var useMaps = typeof array[0] === 'object'
+    if (useMaps) { refSortWithMaps(array, ref) }
+    else { refSortWithObjects(array, ref) }
+    return array
+}
+function refSortWithMaps(array, ref) {
+    var mapping = new Map()
+    array.forEach((val, i) => mapping.set(val, ref[i]))
+    array.sort((a, b) => mapping.get(a) - mapping.get(b))
+}
+function refSortWithObjects(array, ref) {
+    var mapping = {}
+    array.forEach((val, i) => mapping[val] = ref[i])
+    array.sort((a, b) => mapping[a] - mapping[b])
+}
+
+
 
 
 
@@ -58,5 +115,34 @@ export function makeProfileHook(_every, _title) {
         if (state === 'start') start()
         else if (state === 'end') report()
         else add(state)
+    }
+}
+
+
+
+
+// simple thing for reporting time actions/sec
+export function makeThroughputHook(_every, _title) {
+    var title = _title || ''
+    var every = _every || 1
+    var counts = {}
+    var started = performance.now()
+    var iter = 0
+    return function profile_hook(state) {
+        if (state === 'start') return
+        if (state === 'end') {
+            if (++iter < every) return
+            var t = performance.now()
+            console.log(title + '   ' + Object.keys(counts).map(k => {
+                var through = counts[k] / (t - started) * 1000
+                counts[k] = 0
+                return k + ':' + through.toFixed(2) + '   '
+            }).join(''))
+            started = t
+            iter = 0
+        } else {
+            if (!counts[state]) counts[state] = 0
+            counts[state]++
+        }
     }
 }