+++ src/lib/world.js
@@ -1,8 +1,10 @@
 
-var ndHash = require('ndarray-hash')
 var EventEmitter = require('events').EventEmitter
 import Chunk from './chunk'
+import { sortByReferenceArray, loopForTime, numberOfVoxelsInSphere } from './util'
 
+var PROFILE = 0
+var PROFILE_QUEUES = 0
 
 
 export default function (noa, opts) {
@@ -10,8 +12,6 @@ export default function (noa, opts) {
 }
 
 
-var PROFILE = 0
-var PROFILE_QUEUES = 0
 
 
 var defaultOptions = {
@@ -36,7 +36,6 @@ function World(noa, opts) {
     this.noa = noa
     opts = Object.assign({}, defaultOptions, opts)
 
-    this.userData = null
     this.playerChunkLoaded = false
     this.Chunk = Chunk
 
@@ -47,24 +46,21 @@ function World(noa, opts) {
         this.chunkRemoveDistance = this.chunkAddDistance
     }
 
-    // internals
-    this._chunkIDsToAdd = []
-    this._chunkIDsToRemove = []
-    this._chunkIDsInMemory = []
-    this._chunkIDsToCreate = []
-    this._chunkIDsToMesh = []
-    this._chunkIDsToMeshFirst = []
-    this._maxChunksPendingCreation = 20
-    this._maxChunksPendingMeshing = 20
-    this._maxProcessingPerTick = 9 // ms
-    this._maxProcessingPerRender = 5 // ms
+    // settings for tuning worldgen throughput
+    this._maxChunksPendingCreation = 10
+    this._maxChunksPendingMeshing = 10
+    this._maxProcessingPerTick = 9      // ms
+    this._maxProcessingPerRender = 5    // ms
+
+    // set up internal state
+    this._cachedWorldName = ''
+    this._lastPlayerChunkID = ''
+    this._chunkStorage = {}
+    initChunkQueues(this)
+    initChunkStorage(this)
 
     // triggers a short visit to the meshing queue before renders
-    var self = this
-    noa.on('beforeRender', function () { beforeRender(self) })
-
-    // actual chunk storage - hash size hard coded for now
-    this._chunkHash = ndHash([1024, 1024, 1024])
+    noa.on('beforeRender', () => beforeRender(this))
 
     // instantiate coord conversion functions based on the chunk size
     // use bit twiddling if chunk size is a power of 2
@@ -72,48 +68,51 @@ function World(noa, opts) {
     if ((cs & cs - 1) === 0) {
         var shift = Math.log2(cs) | 0
         var mask = (cs - 1) | 0
-        worldCoordToChunkCoord = coord => (coord >> shift) | 0
-        worldCoordToChunkIndex = coord => (coord & mask) | 0
+        this._worldCoordToChunkCoord = coord => (coord >> shift) | 0
+        this._worldCoordToChunkIndex = coord => (coord & mask) | 0
     } else {
-        worldCoordToChunkCoord = coord => Math.floor(coord / cs) | 0
-        worldCoordToChunkIndex = coord => (((coord % cs) + cs) % cs) | 0
+        this._worldCoordToChunkCoord = coord => Math.floor(coord / cs) | 0
+        this._worldCoordToChunkIndex = coord => (((coord % cs) + cs) % cs) | 0
     }
 
 }
 World.prototype = Object.create(EventEmitter.prototype)
 
-var worldCoordToChunkCoord
-var worldCoordToChunkIndex
+
 
 
 
 
 /*
- *   PUBLIC API 
- */
-
-
+ *
+ *
+ *
+ *
+ *                  PUBLIC API 
+ *
+ *
+ *
+ *
+*/
 
 /** @param x,y,z */
 World.prototype.getBlockID = function (x, y, z) {
     var chunk = this._getChunkByCoords(x, y, z)
     if (!chunk) return 0
-
-    var ix = worldCoordToChunkIndex(x)
-    var iy = worldCoordToChunkIndex(y)
-    var iz = worldCoordToChunkIndex(z)
-    return chunk.get(ix, iy, iz)
+    return chunk.get(
+        this._worldCoordToChunkIndex(x),
+        this._worldCoordToChunkIndex(y),
+        this._worldCoordToChunkIndex(z))
 }
 
 /** @param x,y,z */
 World.prototype.getBlockSolidity = function (x, y, z) {
     var chunk = this._getChunkByCoords(x, y, z)
-    if (!chunk) return 0
-
-    var ix = worldCoordToChunkIndex(x)
-    var iy = worldCoordToChunkIndex(y)
-    var iz = worldCoordToChunkIndex(z)
-    return !!chunk.getSolidityAt(ix, iy, iz)
+    if (!chunk) return false
+    return !!chunk.getSolidityAt(
+        this._worldCoordToChunkIndex(x),
+        this._worldCoordToChunkIndex(y),
+        this._worldCoordToChunkIndex(z))
 }
 
 /** @param x,y,z */
@@ -138,22 +137,21 @@ World.prototype.getBlockProperties = function (x, y, z) {
 World.prototype.getBlockObjectMesh = function (x, y, z) {
     var chunk = this._getChunkByCoords(x, y, z)
     if (!chunk) return 0
-
-    var ix = worldCoordToChunkIndex(x)
-    var iy = worldCoordToChunkIndex(y)
-    var iz = worldCoordToChunkIndex(z)
-    return chunk.getObjectMeshAt(ix, iy, iz)
+    return chunk.getObjectMeshAt(
+        this._worldCoordToChunkIndex(x),
+        this._worldCoordToChunkIndex(y),
+        this._worldCoordToChunkIndex(z))
 }
 
 
 /** @param x,y,z */
 World.prototype.setBlockID = function (val, x, y, z) {
-    var i = worldCoordToChunkCoord(x)
-    var j = worldCoordToChunkCoord(y)
-    var k = worldCoordToChunkCoord(z)
-    var ix = worldCoordToChunkIndex(x)
-    var iy = worldCoordToChunkIndex(y)
-    var iz = worldCoordToChunkIndex(z)
+    var i = this._worldCoordToChunkCoord(x)
+    var j = this._worldCoordToChunkCoord(y)
+    var k = this._worldCoordToChunkCoord(z)
+    var ix = this._worldCoordToChunkIndex(x)
+    var iy = this._worldCoordToChunkIndex(y)
+    var iz = this._worldCoordToChunkIndex(z)
 
     // if update is on chunk border, update neighbor's padding data too
     _updateChunkAndBorders(this, i, j, k, this.chunkSize, ix, iy, iz, val)
@@ -175,59 +173,6 @@ World.prototype.isBoxUnobstructed = function (box) {
 }
 
 
-
-
-
-World.prototype.tick = function () {
-    profile_hook('start')
-
-    // check player position and needed/unneeded chunks
-    var pos = getPlayerChunkCoords(this)
-    var chunkID = getChunkID(pos[0], pos[1], pos[2])
-    if (chunkID != this._lastPlayerChunkID) {
-        this.emit('playerEnteredChunk', pos[0], pos[1], pos[2])
-        buildChunkAddQueue(this, pos[0], pos[1], pos[2])
-        buildChunkRemoveQueue(this, pos[0], pos[1], pos[2])
-    }
-    this._lastPlayerChunkID = chunkID
-    profile_hook('build queues')
-
-    // process (create or mesh) some chunks. If fast enough, do several
-    profile_queues(this, 'start')
-    var cutoff = performance.now() + this._maxProcessingPerTick
-    var done = false
-    while (!done && (performance.now() < cutoff)) {
-        var d1 = processMeshingQueues(this, false)
-        var d2 = processChunkQueues(this)
-        if (!d2) d2 = processChunkQueues(this)
-        done = d1 && d2
-    }
-    profile_queues(this, 'end')
-
-
-    // track whether the player's local chunk is loaded and ready or not
-    var pChunk = getChunk(this, pos[0], pos[1], pos[2])
-    var okay = !!(pChunk && pChunk.isGenerated && !pChunk.isInvalid)
-    this.playerChunkLoaded = okay
-
-    profile_hook('end')
-}
-
-
-
-function beforeRender(self) {
-    // on render, quickly process the high-priority meshing queue
-    // to help avoid flashes of background while neighboring chunks update
-    var cutoff = performance.now() + self._maxProcessingPerRender
-    var done = false
-    while (!done && (performance.now() < cutoff)) {
-        done = processMeshingQueues(self, true)
-    }
-}
-
-
-
-
 /** client should call this after creating a chunk's worth of data (as an ndarray)  
  * If userData is passed in it will be attached to the chunk
  * @param id
@@ -235,181 +180,275 @@ function beforeRender(self) {
  * @param userData
  */
 World.prototype.setChunkData = function (id, array, userData) {
-    profile_queues(this, 'received')
-    var arr = parseChunkID(id)
-    var chunk = getChunk(this, arr[0], arr[1], arr[2])
-    // ignore if chunk was invalidated while being prepared
-    if (!chunk || chunk.isInvalid) return
-    chunk.array = array
-    if (userData) chunk.userData = userData
-    chunk.initData()
-    enqueueID(id, this._chunkIDsInMemory)
-    unenqueueID(id, this._chunkIDsToCreate)
-
-    // chunk can now be meshed...
-    this.noa.rendering.prepareChunkForRendering(chunk)
-    enqueueID(id, this._chunkIDsToMesh)
-    this.emit('chunkAdded', chunk)
+    setChunkData(this, id, array, userData)
 }
 
 
-
-
-/*
- * Calling this causes all world chunks to get unloaded and recreated 
- * (after receiving new world data from the client). This is useful when
- * you're teleporting the player to a new world, e.g.
+/** Tells noa to discard all chunks in memory and request new data from client.
+ * Clients *probably* shouldn't need this API anymore - to change between 
+ * multiple sets of world data use `noa.worldName`.
  */
 World.prototype.invalidateAllChunks = function () {
-    var toInval = this._chunkIDsInMemory.concat(this._chunkIDsToCreate)
-    toInval.forEach(id => {
-        var loc = parseChunkID(id)
-        var chunk = getChunk(this, loc[0], loc[1], loc[2])
-        chunk.isInvalid = true
-    })
-    // this causes chunk queues to get rebuilt next tick
-    this._lastPlayerChunkID += '.'
+    markAllChunksForRemoval(this)
 }
 
 
 
-// debugging
-World.prototype.report = function () {
-    console.log('World report - playerChunkLoaded: ', this.playerChunkLoaded)
-    _report(this, '  to add     ', this._chunkIDsToAdd)
-    _report(this, '  to remove: ', this._chunkIDsToRemove)
-    _report(this, '  in memory: ', this._chunkIDsInMemory, true)
-    _report(this, '  creating:  ', this._chunkIDsToCreate)
-    _report(this, '  meshing:   ', this._chunkIDsToMesh.concat(this._chunkIDsToMeshFirst))
-}
-
-function _report(world, name, arr, ext) {
-    var ct = 0,
-        full = 0,
-        empty = 0
-    for (var id of arr) {
-        if (id.size) {
-            if (id.isInvalid) ct++
-            continue
-        }
-        var loc = parseChunkID(id)
-        var chunk = getChunk(world, loc[0], loc[1], loc[2])
-        if (chunk.isInvalid) ct++
-        if (chunk.isFull) full++
-        if (chunk.isEmpty) empty++
-    }
-    var len = (arr.length + '        ').substr(0, 6)
-    var es = (ext) ? [', ', full, ' full, ', empty, ' empty'].join('') : ''
-    console.log(name, len, ct, 'invalid' + es)
-}
 
 
 
 
 /*
- *
- *
- *            INTERNALS
- *
- *
- */
+ * 
+ * 
+ * 
+ *                  internals:
+ * 
+ *          tick functions that process queues and trigger events
+ * 
+ * 
+ * 
+*/
 
 
-// canonical string ID handling for the i,j,k-th chunk
+World.prototype.tick = function () {
+    var tickStartTime = performance.now()
+
+    // if world has changed, invalidate everything
+    if (this._cachedWorldName !== this.noa.worldName) {
+        this.invalidateAllChunks()
+        this._cachedWorldName = this.noa.worldName
+    }
+
+    // current player chunk changed since last tick?
+    var pos = getPlayerChunkCoords(this)
+    var chunkID = getChunkID(pos[0], pos[1], pos[2])
+    var changedChunks = (chunkID != this._lastPlayerChunkID)
+    if (changedChunks) {
+        this.emit('playerEnteredChunk', pos[0], pos[1], pos[2])
+        this._lastPlayerChunkID = chunkID
+    }
+    profile_hook('start')
+    profile_queues_hook('start')
+
+    // possibly scan for chunks to add/remove
+    if (changedChunks) {
+        findDistantChunksToRemove(this, pos[0], pos[1], pos[2])
+        profile_hook('remQueue')
+    }
+    var numChunks = numberOfVoxelsInSphere(this.chunkAddDistance)
+    if (changedChunks || this._chunkIDsKnown.length < numChunks) {
+        findNewChunksInRange(this, pos[0], pos[1], pos[2])
+        profile_hook('addQueue')
+    }
+
+    // process (create or mesh) some chunks, up to max iteration time
+    loopForTime(this._maxProcessingPerTick, () => {
+        var done = processRequestQueue(this)
+        profile_hook('requests')
+        done = done && processRemoveQueue(this)
+        profile_hook('removes')
+        done = done && processMeshingQueue(this, false)
+        profile_hook('meshes')
+        return done
+    }, tickStartTime)
+
+    // track whether the player's local chunk is loaded and ready or not
+    var pChunk = this._getChunk(pos[0], pos[1], pos[2])
+    this.playerChunkLoaded = !!pChunk
+
+    profile_queues_hook('end')
+    profile_hook('end')
+}
+
+
+
+function beforeRender(world) {
+    // on render, quickly process the high-priority meshing queue
+    // to help avoid flashes of background while neighboring chunks update
+    loopForTime(world._maxProcessingPerRender, () => {
+        return processMeshingQueue(world, true)
+    })
+}
+
+
+
+
+
+
+
+
+/*
+ * 
+ * 
+ * 
+ *              chunk IDs, storage, and lookup/retrieval
+ * 
+ * 
+ * 
+*/
+
 function getChunkID(i, j, k) {
+    // chunk coords -> canonical string ID
     return i + '|' + j + '|' + k
 }
 
 function parseChunkID(id) {
+    // chunk ID -> coords
     var arr = id.split('|')
     return [parseInt(arr[0]), parseInt(arr[1]), parseInt(arr[2])]
 }
 
-// canonical functions to store/retrieve a chunk held in memory
-function getChunk(world, i, j, k) {
-    var mi = (i | 0) & 1023
-    var mj = (j | 0) & 1023
-    var mk = (k | 0) & 1023
-    return world._chunkHash.get(mi, mj, mk)
+function initChunkStorage(world) {
+    // var chunkHash = ndHash([1024, 1024, 1024])
+    world._getChunk = (i, j, k) => {
+        var id = getChunkID(i, j, k)
+        return world._chunkStorage[id] || null
+    }
+    world._setChunk = (i, j, k, value) => {
+        var id = getChunkID(i, j, k)
+        if (value) {
+            world._chunkStorage[id] = value
+        } else {
+            delete world._chunkStorage[id]
+        }
+    }
+    // chunk accessor for internal use
+    world._getChunkByCoords = function (x, y, z) {
+        var i = world._worldCoordToChunkCoord(x)
+        var j = world._worldCoordToChunkCoord(y)
+        var k = world._worldCoordToChunkCoord(z)
+        return world._getChunk(i, j, k)
+    }
 }
-
-function setChunk(world, i, j, k, value) {
-    var mi = (i | 0) & 1023
-    var mj = (j | 0) & 1023
-    var mk = (k | 0) & 1023
-    world._chunkHash.set(mi, mj, mk, value)
-}
-
-
 
 function getPlayerChunkCoords(world) {
     var pos = world.noa.entities.getPosition(world.noa.playerEntity)
-    var i = worldCoordToChunkCoord(pos[0])
-    var j = worldCoordToChunkCoord(pos[1])
-    var k = worldCoordToChunkCoord(pos[2])
+    var i = world._worldCoordToChunkCoord(pos[0])
+    var j = world._worldCoordToChunkCoord(pos[1])
+    var k = world._worldCoordToChunkCoord(pos[2])
     return [i, j, k]
 }
 
 
-// for internal use
-World.prototype._getChunkByCoords = function (x, y, z) {
-    var i = worldCoordToChunkCoord(x)
-    var j = worldCoordToChunkCoord(y)
-    var k = worldCoordToChunkCoord(z)
-    return getChunk(this, i, j, k)
+
+
+
+
+
+
+/*
+ * 
+ * 
+ * 
+ *              chunk queues and queue processing
+ * 
+ * 
+ * 
+*/
+
+
+function initChunkQueues(world) {
+    world._chunkIDsKnown = []       // all chunks existing in any queue
+    world._chunkIDsToRequest = []   // not yet requested from client
+    world._chunkIDsPending = []     // requested, awaiting creation
+    world._chunkIDsToMesh = []      // created but not yet meshed
+    world._chunkIDsToMeshFirst = [] // priority meshing queue
+    world._chunkIDsToRemove = []    // chunks awaiting disposal
 }
 
 
+// process neighborhood chunks, add missing ones to "toRequest" and "inMemory"
+function findNewChunksInRange(world, ci, cj, ck) {
+    var add = Math.ceil(world.chunkAddDistance)
+    var addDistSq = world.chunkAddDistance * world.chunkAddDistance
+    var known = world._chunkIDsKnown
+    var toRequest = world._chunkIDsToRequest
+    // search all nearby chunk locations
+    for (var i = ci - add; i <= ci + add; ++i) {
+        for (var j = cj - add; j <= cj + add; ++j) {
+            for (var k = ck - add; k <= ck + add; ++k) {
+                var id = getChunkID(i, j, k)
+                if (known.includes(id)) continue
+                var di = i - ci
+                var dj = j - cj
+                var dk = k - ck
+                var distSq = di * di + dj * dj + dk * dk
+                if (distSq > addDistSq) continue
+                enqueueID(id, known)
+                enqueueID(id, toRequest)
+            }
+        }
+    }
+    sortChunkIDQueue(world, toRequest)
+}
+
+
+// rebuild queue of chunks to be removed from around (ci,cj,ck)
+function findDistantChunksToRemove(world, ci, cj, ck) {
+    var remDistSq = world.chunkRemoveDistance * world.chunkRemoveDistance
+    var toRemove = world._chunkIDsToRemove
+    world._chunkIDsKnown.forEach(id => {
+        if (toRemove.includes(id)) return
+        var loc = parseChunkID(id)
+        var di = loc[0] - ci
+        var dj = loc[1] - cj
+        var dk = loc[2] - ck
+        var distSq = di * di + dj * dj + dk * dk
+        if (distSq < remDistSq) return
+        // flag chunk for removal and remove it from work queues
+        enqueueID(id, world._chunkIDsToRemove)
+        unenqueueID(id, world._chunkIDsToRequest)
+        unenqueueID(id, world._chunkIDsToMesh)
+        unenqueueID(id, world._chunkIDsToMeshFirst)
+    })
+    sortChunkIDQueue(world, toRemove)
+}
+
+
+// when current world changes - empty work queues and mark all for removal
+function markAllChunksForRemoval(world) {
+    world._chunkIDsToRemove = world._chunkIDsKnown.slice()
+    world._chunkIDsToRequest.length = 0
+    world._chunkIDsToMesh.length = 0
+    world._chunkIDsToMeshFirst.length = 0
+    sortChunkIDQueue(world, world._chunkIDsToRemove)
+}
 
 
 // run through chunk tracking queues looking for work to do next
-function processChunkQueues(self) {
-    var done = true
-    // both queues are sorted by ascending distance
-    if (self._chunkIDsToRemove.length) {
-        var remove = parseChunkID(self._chunkIDsToRemove.pop())
-        removeChunk(self, remove[0], remove[1], remove[2])
-        profile_queues(self, 'removed')
-        profile_hook('removed')
-        done = false
-    }
-    if (self._chunkIDsToCreate.length >= self._maxChunksPendingCreation) return done
-    // if (self._chunkIDsToMesh.length >= self._maxChunksPendingMeshing) return done
-    if (self._chunkIDsToAdd.length) {
-        var id = self._chunkIDsToAdd.shift()
-        requestNewChunk(self, id)
-        profile_hook('requested')
-        profile_queues(self, 'requested')
-        done = false
-    }
-    return done
+function processRequestQueue(world) {
+    var queue = world._chunkIDsToRequest
+    if (queue.length === 0) return true
+    // skip if too many outstanding requests, or if meshing queue is full
+    var pending = world._chunkIDsPending.length
+    var toMesh = world._chunkIDsToMesh.length
+    if (pending >= world._maxChunksPendingCreation) return true
+    if (toMesh >= world._maxChunksPendingMeshing) return true
+    var id = queue.shift()
+    requestNewChunk(world, id)
+    return (queue.length === 0)
+}
+
+
+function processRemoveQueue(world) {
+    var queue = world._chunkIDsToRemove
+    if (queue.length === 0) return true
+    removeChunk(world, queue.shift())
+    return (queue.length === 0)
 }
 
 
 // similar to above but for chunks waiting to be meshed
-function processMeshingQueues(self, firstOnly) {
-    var id
-    if (self._chunkIDsToMeshFirst.length) {
-        id = self._chunkIDsToMeshFirst.pop()
-    } else if (firstOnly) {
-        return true
-    } else if (self._chunkIDsToMesh.length) {
-        id = self._chunkIDsToMesh.pop()
-    } else return true
+function processMeshingQueue(world, firstOnly) {
+    var queue = world._chunkIDsToMeshFirst
+    if (queue.length === 0 && !firstOnly) queue = world._chunkIDsToMesh
+    if (queue.length === 0) return true
 
-    var arr = parseChunkID(id)
-    var chunk = getChunk(self, arr[0], arr[1], arr[2])
-    if (chunk.isInvalid) return
-    if (!chunk.isGenerated) {
-        // client code triggered a remesh too early, requeue it
-        self._chunkIDsToMesh.unshift(id)
-        return
-    }
-    chunk.updateMeshes()
-
-    profile_queues(self, 'meshed')
-    profile_hook('meshed')
-    return false
+    var id = queue.shift()
+    if (world._chunkIDsToRemove.includes(id)) return
+    var chunk = world._chunkStorage[id]
+    if (chunk) doChunkRemesh(world, chunk)
 }
 
 
@@ -420,40 +459,80 @@ function processMeshingQueues(self, firstOnly) {
 
 
 
-// make a new chunk and emit an event for it to be populated with world data
+/*
+ * 
+ * 
+ * 
+ *              chunk lifecycle - create / set / remove / modify
+ * 
+ * 
+ * 
+*/
+
+
+// create chunk object and request voxel data from client
 function requestNewChunk(world, id) {
     var pos = parseChunkID(id)
     var i = pos[0]
     var j = pos[1]
     var k = pos[2]
     var size = world.chunkSize
-    var chunk = new Chunk(world.noa, id, i, j, k, size)
-    setChunk(world, i, j, k, chunk)
+    var dataArr = Chunk.createVoxelArray(world.chunkSize)
+    var worldName = world.noa.worldName
+    var requestID = [i, j, k, worldName].join('|')
     var x = i * size - 1
     var y = j * size - 1
     var z = k * size - 1
-    enqueueID(id, world._chunkIDsToCreate)
-    world.emit('worldDataNeeded', id, chunk.array, x, y, z)
+    enqueueID(id, world._chunkIDsPending)
+    world.emit('worldDataNeeded', requestID, dataArr, x, y, z, worldName)
+    profile_queues_hook('request')
 }
 
+// called when client sets a chunk's voxel data
+// If userData is passed in it will be attached to the chunk
+function setChunkData(world, reqID, array, userData) {
+    var arr = reqID.split('|')
+    var i = arr.shift()
+    var j = arr.shift()
+    var k = arr.shift()
+    var worldName = arr.join('|')
+    var id = getChunkID(i, j, k)
+    unenqueueID(id, world._chunkIDsPending)
+    // discard data if it's for a world that's no longer current
+    if (worldName !== world.noa.worldName) return
+    // discard if chunk is no longer needed
+    if (!world._chunkIDsKnown.includes(id)) return
+    if (world._chunkIDsToRemove.includes(id)) return
+    // all good, create and initialize the chunk
+    var size = world.chunkSize
+    var chunk = new Chunk(world.noa, id, i, j, k, size, array)
+    world._setChunk(i, j, k, chunk)
+    chunk.requestID = reqID
+    chunk.userData = userData
+    // chunk can now be meshed...
+    world.noa.rendering.prepareChunkForRendering(chunk)
+    enqueueID(id, world._chunkIDsToMesh)
+    world.emit('chunkAdded', chunk)
+    profile_queues_hook('receive')
+}
 
 
 
 // remove a chunk that wound up in the remove queue
-function removeChunk(world, i, j, k) {
-    var chunk = getChunk(world, i, j, k)
-    world.emit('chunkBeingRemoved', chunk.id, chunk.array, chunk.userData)
-    world.noa.rendering.disposeChunkForRendering(chunk)
-    chunk.dispose()
-    setChunk(world, i, j, k, 0)
-    unenqueueID(chunk.id, world._chunkIDsInMemory)
-    unenqueueID(chunk.id, world._chunkIDsToMesh)
-    unenqueueID(chunk.id, world._chunkIDsToMeshFirst)
-    // when removing a chunk because it was invalid, arrange for chunk queues to get rebuilt
-    if (chunk.isInvalid) world._lastPlayerChunkID += '.'
+function removeChunk(world, id) {
+    var loc = parseChunkID(id)
+    var chunk = world._getChunk(loc[0], loc[1], loc[2])
+    if (chunk) {
+        world.emit('chunkBeingRemoved', chunk.requestID, chunk.array, chunk.userData)
+        world.noa.rendering.disposeChunkForRendering(chunk)
+        chunk.dispose()
+        profile_queues_hook('dispose')
+    }
+    world._setChunk(loc[0], loc[1], loc[2], null)
+    unenqueueID(id, world._chunkIDsKnown)
+    unenqueueID(id, world._chunkIDsToMesh)
+    unenqueueID(id, world._chunkIDsToMeshFirst)
 }
-
-
 
 
 
@@ -485,13 +564,9 @@ function _updateChunkAndBorders(world, i, j, k, size, x, y, z, val) {
         }
     }
 }
-
-
-
 // internal function to modify a chunk's block
-
 function _modifyBlockData(world, i, j, k, x, y, z, val, isPadding) {
-    var chunk = getChunk(world, i, j, k)
+    var chunk = world._getChunk(i, j, k)
     if (!chunk) return
     chunk.set(x, y, z, val, isPadding)
     enqueueID(chunk.id, world._chunkIDsToMeshFirst)
@@ -500,74 +575,33 @@ function _modifyBlockData(world, i, j, k, x, y, z, val, isPadding) {
 
 
 
-
-// rebuild queue of chunks to be added around (ci,cj,ck)
-function buildChunkAddQueue(world, ci, cj, ck) {
-    var add = Math.ceil(world.chunkAddDistance)
-    var pending = world._chunkIDsToCreate
-    var queue = []
-    var distArr = []
-
-    var addDistSq = world.chunkAddDistance * world.chunkAddDistance
-    for (var i = ci - add; i <= ci + add; ++i) {
-        for (var j = cj - add; j <= cj + add; ++j) {
-            for (var k = ck - add; k <= ck + add; ++k) {
-                var di = i - ci
-                var dj = j - cj
-                var dk = k - ck
-                var horizDistSq = di * di + dk * dk
-                var totalDistSq = horizDistSq + dj * dj
-                if (totalDistSq > addDistSq) continue
-
-                if (getChunk(world, i, j, k)) continue
-                var id = getChunkID(i, j, k)
-                if (pending.indexOf(id) > -1) continue
-                queue.push(id)
-                distArr.push(horizDistSq + Math.abs(dj))
-            }
-        }
-    }
-    world._chunkIDsToAdd = sortByReferenceArray(queue, distArr)
-}
-
-
-// rebuild queue of chunks to be removed from around (ci,cj,ck)
-function buildChunkRemoveQueue(world, ci, cj, ck) {
-    var remDistSq = world.chunkRemoveDistance * world.chunkRemoveDistance
-    var list = world._chunkIDsInMemory
-    var queue = []
-    var distArr = []
-
-    for (var i = 0; i < list.length; i++) {
-        var id = list[i]
-        var loc = parseChunkID(id)
-        var di = loc[0] - ci
-        var dj = loc[1] - cj
-        var dk = loc[2] - ck
-        var distSq = di * di + dj * dj + dk * dk
-        if (distSq < remDistSq) {
-            var chunk = getChunk(world, loc[0], loc[1], loc[2])
-            if (!chunk.isInvalid) continue
-            distSq *= -1 // rig sort so that invalidated chunks get removed first
-        }
-        queue.push(id)
-        distArr.push(distSq)
-    }
-    world._chunkIDsToRemove = sortByReferenceArray(queue, distArr)
-}
-
-
-
-// sorts [A, B, C] and [3, 1, 2] into [B, C, A]
-function sortByReferenceArray(data, ref) {
-    var ind = Object.keys(ref)
-    ind.sort((i, j) => ref[i] - ref[j])
-    return ind.map(i => data[i])
+function doChunkRemesh(world, chunk) {
+    unenqueueID(chunk.id, world._chunkIDsToMesh)
+    unenqueueID(chunk.id, world._chunkIDsToMeshFirst)
+    chunk.updateMeshes()
+    profile_queues_hook('mesh')
 }
 
 
 
 
+
+
+
+
+
+
+
+
+/*
+ * 
+ * 
+ * 
+ *          misc helpers and implementation functions
+ * 
+ * 
+ * 
+*/
 
 // uniquely enqueue a string id into an array of them
 function enqueueID(id, queue) {
@@ -583,76 +617,69 @@ function unenqueueID(id, queue) {
 }
 
 
-
-
-
-var profile_queues = function (w, s) {}
-if (PROFILE_QUEUES)(function () {
-    var every = 100
-    var iter = 0
-    var t, nrem, nreq, totalrec, nmesh
-    var reqcts, remcts, meshcts
-    var qadd, qrem, qmem, qgen, qmesh
-    profile_queues = function (world, state) {
-        if (state === 'start') {
-            if (iter === 0) {
-                t = performance.now()
-                qadd = qrem = qmem = qgen = qmesh = 0
-                totalrec = 0
-                remcts = []
-                reqcts = []
-                meshcts = []
-            }
-            iter++
-            nrem = nreq = nmesh = 0
-        } else if (state === 'removed') {
-            nrem++
-        } else if (state === 'received') {
-            totalrec++
-        } else if (state === 'requested') {
-            nreq++
-        } else if (state === 'meshed') {
-            nmesh++
-        } else if (state === 'end') {
-            // counts for frames that were fully worked
-            if (world._chunkIDsToAdd.length) reqcts.push(nreq)
-            if (world._chunkIDsToRemove.length) remcts.push(nrem)
-            if (world._chunkIDsToMesh.length + world._chunkIDsToMeshFirst.length) meshcts.push(nmesh)
-            // avg queue sizes
-            qadd += world._chunkIDsToAdd.length
-            qrem += world._chunkIDsToRemove.length
-            qmem += world._chunkIDsInMemory.length
-            qgen += world._chunkIDsToCreate.length
-            qmesh += world._chunkIDsToMesh.length + world._chunkIDsToMeshFirst.length
-            // on end
-            if (iter === every) {
-                var dt = (performance.now() - t) / 1000
-                console.log('world chunk queues:',
-                    'made', rnd(totalrec / dt), 'cps',
-                    '- avg queuelen: ',
-                    'add', qadd / every,
-                    'rem', qrem / every,
-                    'mem', qmem / every,
-                    'gen', qgen / every,
-                    'mesh', qmesh / every,
-                    '- work/frame: ',
-                    'req', rnd(reqcts.reduce(sum, 0) / reqcts.length),
-                    'rem', rnd(remcts.reduce(sum, 0) / remcts.length),
-                    'mesh', rnd(meshcts.reduce(sum, 0) / meshcts.length)
-                )
-                iter = 0
-            }
-        }
-    }
-    var sum = function (num, prev) { return num + prev }
-    var rnd = function (n) { return Math.round(n * 10) / 10 }
-})()
+// sorts a queue of chunk IDs by distance from player (ascending)
+function sortChunkIDQueue(world, queue) {
+    var loc = getPlayerChunkCoords(world)
+    var dists = queue.map(id => {
+        var pos = parseChunkID(id)
+        var dx = pos[0] - loc[0]
+        var dy = pos[1] - loc[1]
+        var dz = pos[2] - loc[2]
+        // bias towards keeping verticals together for now
+        return 3 * (dx * dx + dz * dz) + Math.abs(dy)
+    })
+    sortByReferenceArray(queue, dists)
+}
 
 
 
 
 
 
-import { makeProfileHook } from './util'
+
+
+
+/*
+ * 
+ * 
+ * 
+ * 
+ *                  debugging
+ * 
+ * 
+ * 
+ * 
+*/
+
+World.prototype.report = function () {
+    console.log('World report - playerChunkLoaded: ', this.playerChunkLoaded)
+    _report(this, '  known:     ', this._chunkIDsKnown)
+    _report(this, '  to request:', this._chunkIDsToRequest)
+    _report(this, '  to remove: ', this._chunkIDsToRemove)
+    _report(this, '  creating:  ', this._chunkIDsPending)
+    _report(this, '  to mesh:   ', this._chunkIDsToMesh.concat(this._chunkIDsToMeshFirst))
+}
+
+function _report(world, name, arr, ext) {
+    var full = 0,
+        empty = 0,
+        exist = 0
+    arr.forEach(id => {
+        var chunk = world._chunkStorage[id]
+        if (chunk) exist++
+        if (chunk && chunk.isFull) full++
+        if (chunk && chunk.isEmpty) empty++
+    })
+    var out = arr.length.toString().padEnd(8)
+    out += ('exist: ' + exist).padEnd(12)
+    out += ('full: ' + full).padEnd(12)
+    out += ('empty: ' + empty).padEnd(12)
+    console.log(name, out)
+}
+
+
+import { makeProfileHook, makeThroughputHook } from './util'
 var profile_hook = (PROFILE) ?
-    makeProfileHook(200, 'world ticks') : () => {}
+    makeProfileHook(100, 'world ticks:') : () => { }
+var profile_queues_hook = (PROFILE_QUEUES) ?
+    makeThroughputHook(100, 'chunks/sec:') : () => { }